+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 
 ## Additions
 - Added config `BoardCellFont` that allows to change the font of the board's cells for better readability
+- Added config `BoardPosition` that allows to change where the board is located
+- Added config `BoardScale` that allows to change how big the board is shown
+- Added config `BoardOpacity` that allows to change the opacity of the board
 - Added condition `has_obtained_quill` with 3 simple goals
 - Added condition `has_obtained_map` with 28 simple goals
 - Added condition `has_completed_wish` with 54 simple goals